# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lander Dispatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to OIDC
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect on success
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
            default: /
      responses:
        '302':
          description: Redirect to OIDC logout

  # Dashboard
...
  /brokers:
    get:
      tags: [Brokers]
      operationId: listBrokers
      summary: List brokers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated broker list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerListResponse'
    post:
      tags: [Brokers]
      operationId: createBroker
      summary: Create a broker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Broker'
  /brokers/{brokerId}:
    get:
      tags: [Brokers]
      operationId: getBroker
      summary: Get a broker by ID
      parameters:
        - name: brokerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Broker detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Broker'
    patch:
      tags: [Brokers]
      operationId: updateBroker
      summary: Update a broker
      parameters:
        - name: brokerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Broker'
    delete:
      tags: [Brokers]
      operationId: deleteBroker
      summary: Delete a broker
      parameters:
        - name: brokerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # Loads
...
  /accounting/profit-loss:
    get:
      tags: [ProfitLoss]
      operationId: getProfitLoss
      summary: Get profit and loss report
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            default: this_month
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: P&L report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfitLossReport'

  # Settings
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    # Pagination
    PaginationMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # Dashboard
...
    Broker:
      type: object
      required: [id, companyName, status, createdAt]
      properties:
        id:
          type: string
        companyName:
          type: string
        brokerType:
          type: [string, 'null']
        website:
          type: [string, 'null']
        mcNumber:
          type: [string, 'null']
        usdotNumber:
          type: [string, 'null']
        primaryContact:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        email:
          type: [string, 'null']
        lastContact:
          type: [string, 'null']
          format: date
        freightTypes:
          type: array
          items:
            type: string
        coverage:
          type: [string, 'null']
        selectedStates:
          type: array
          items:
            type: string
        paymentTerms:
          type: [string, 'null']
        paymentDays:
          type: [integer, 'null']
        quickPay:
          type: [boolean, 'null']
        quickPayFee:
          type: [number, 'null']
        factoringAccepted:
          type: [string, 'null']
        onboardingStatus:
          type: [string, 'null']
        rating:
          type: [number, 'null']
        status:
          type: string
        priority:
          type: [string, 'null']
        tags:
          type: array
          items:
            type: string
        notes:
          type: [string, 'null']
        lastLoadDate:
          type: [string, 'null']
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    BrokerInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
          minLength: 1
        brokerType:
          type: string
        website:
          type: string
        mcNumber:
          type: string
        usdotNumber:
          type: string
        primaryContact:
          type: string
        phone:
          type: string
        email:
          type: string
        lastContact:
          type: string
        freightTypes:
          type: array
          items:
            type: string
        coverage:
          type: string
        selectedStates:
          type: array
          items:
            type: string
        paymentTerms:
          type: string
        paymentDays:
          type: integer
        quickPay:
          type: boolean
        quickPayFee:
          type: number
        factoringAccepted:
          type: string
        onboardingStatus:
          type: string
        rating:
          type: number
        status:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    BrokerUpdate:
      type: object
      properties:
        companyName:
          type: string
          minLength: 1
        brokerType:
          type: string
        website:
          type: string
        mcNumber:
          type: string
        usdotNumber:
          type: string
        primaryContact:
          type: string
        phone:
          type: string
        email:
          type: string
        lastContact:
          type: string
        freightTypes:
          type: array
          items:
            type: string
        coverage:
          type: string
        selectedStates:
          type: array
          items:
            type: string
        paymentTerms:
          type: string
        paymentDays:
          type: integer
        quickPay:
          type: boolean
        quickPayFee:
          type: number
        factoringAccepted:
          type: string
        onboardingStatus:
          type: string
        rating:
          type: number
        status:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    BrokerListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Broker'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Load
...
    PnlCategoryLine:
      type: object
      required: [category, amount]
      properties:
        category:
          type: string
        amount:
          type: number
    ProfitLossReport:
      type: object
      required: [period, startDate, endDate, totalRevenue, totalExpenses, netProfit, netProfitMargin]
      properties:
        period:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        netProfitMargin:
          type: number
        revenueLines:
          type: array
          items:
            $ref: '#/components/schemas/PnlCategoryLine'
        expenseLines:
          type: array
          items:
            $ref: '#/components/schemas/PnlCategoryLine'
        previousPeriodRevenue:
          type: [number, 'null']
        previousPeriodExpenses:
          type: [number, 'null']
        previousPeriodNetProfit:
          type: [number, 'null']

    # Settings
...